Updates Auto-Generated Configuration for Sequelize CLI
When Sequelize CLI is initialized using the command sequelize init,
a configuration file is generated. The file initially is in JSON
format, and attempting to run a migration with the file in its
original form was failing.

In addition, I wanted to be able to use environment variables in
the configuration file, so I created the initialization such that
it generates a JavaScript file (instead of JSON).

To work around the errors that were being thrown while attempting
to run migrations, I converted the generated JSON object to
JavaScript and added `module.exports =` at the beginning of the
file.

Your development environment should be configured to include:

* An environment variable named POSTGRES_USER
* An environment variable named SIREN_DB_DEV

A null password is assumed.