Initialize Database Associations
When creating models in Sequelize, we have the ability to add
associations in the generated files. These associations are not
automatically synchronized in the database when a migration is
run.

This file was created as a workaround.

After running a migration for the first time, execute the command:

node middleware/initialize-db.js

This will create the necessary join tables in the database.