Include Pending Migrations
Unlike Rails, the Sequelize CLI does not allow you to specify that
you want to run a particular migration or a set of migrations up to
a certain timestamp. Because the following migrations depend upon the
associations being present, and Sequelize CLI does not appear to do
this out of the box, middleware has been included in the repo that
handles this process.

Once the initial migration is run and the middleware is executed,
move these migrations to the default migrations folder and run the
sequelize db:migrate command