Finalize PodcastConstroller Subscribe Function
Updates the Subscribe function such that when a user subscribes
to a podcast, the podcast entry is added to the Podcasts model
(if it does not exist) and the UserPodcats model is updated with
the UserId and Podcast Id. Next, the Episodes are added to the Episodes
model. Then, we query for the latest 10 podcasts and insert them
into the UserEpisodes Table.

Other minor changes:

* Moves secret.json into the Config folder
* Refactors files the use this value