Update Configuration for Deployment
Adds two configuration variables to config.js that can be toggled
to determine whether or not the database should force sync which
drops and recreates all of the tables.

After schema changes, go into config/db.js and change the
dbChangeSchema variable to true.

At this point, run the seeder and manually run the SQL statement to
create the Inbox view.