Subscription Controller Refactoring
Adds return statements before all nested sequelize calls (initially,
I was only adding the return statement if the next .then() in the chain
needed access to the data that was being returned from the prior
promise).

Adds catch statement that resolves the Promise when the Episode
subscriber arrives at a podcast that it can't process (in one example,
it was missing episode.enclosure.url). (subscriptionHelpers.js line 81)